requirements.txt
asarPy==1.0.1
asgiref==3.11.1
bleach==6.3.0
boto3==1.42.64
botocore==1.42.64
Django==6.0.3
django-markdownify==0.9.6
django-storages==1.14.6
django-taggit==6.1.0
gunicorn==25.1.0
jmespath==1.1.0
Markdown==3.10.2
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tinycss2==1.4.0
urllib3==2.6.3
webencodings==0.5.1
whitenoise==6.12.0
